Criar permissões para `Equipamentos\Cadastro\Modelo`

+++ app/Http/Controllers/Admin/Equipamentos/Cadastro/ModeloController.php
@@ -8,11 +8,13 @@ use App\Http\Requests\Admin\Equipamentos\Cadastro\ModeloRequest;
 use App\Models\Equipamentos\Cadastro\Modelo;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Gate;
 
 class ModeloController extends Controller
 {
     public function inicio()
     {
+        Gate::authorize('ver', Modelo::class);
         $modelos = Modelo::with(['marca'])->paginate(10);
 
         $statusCadastro = StatusCadastro::toArray();
@@ -22,6 +24,7 @@ class ModeloController extends Controller
 
     public function criar()
     {
+        Gate::authorize('criar', Modelo::class);
         $statusCadastro = StatusCadastro::toArray();
 
         return Inertia::render('Admin/Equipamentos/Cadastro/Modelo/Criar', compact('statusCadastro'));
@@ -29,6 +32,7 @@ class ModeloController extends Controller
 
     public function salvar(ModeloRequest $request)
     {
+        Gate::authorize('criar', Modelo::class);
         Modelo::create($request->all());
 
         return redirect()->route('admin.modelos');
@@ -36,6 +40,7 @@ class ModeloController extends Controller
 
     public function editar(int $id)
     {
+        Gate::authorize('editar', Modelo::class);
         $modelo = Modelo::with('marca')->findOrFail($id);
 
         $statusCadastro = StatusCadastro::toArray();
@@ -48,6 +53,7 @@ class ModeloController extends Controller
 
     public function atualizar(ModeloRequest $request, int $id)
     {
+        Gate::authorize('editar', Modelo::class);
         Modelo::findOrFail($id)->update($request->all());
 
         return redirect()->route('admin.modelos');
@@ -55,6 +61,7 @@ class ModeloController extends Controller
 
     public function excluir(int $id)
     {
+        Gate::authorize('excluir', Modelo::class);
         Modelo::findOrFail($id)->delete();
 
         return redirect()->route('admin.modelos');
@@ -62,6 +69,7 @@ class ModeloController extends Controller
 
     public function pesquisar(Request $request, ?int $marcaId = null)
     {
+        Gate::authorize('ver', Modelo::class);
         $modelos = Modelo::select('id', 'nome as texto')
             ->where(function ($query) use ($request) {
                 $query
@@ -79,6 +87,7 @@ class ModeloController extends Controller
 
     public function salvarAjax(ModeloRequest $request)
     {
+        Gate::authorize('criar', Modelo::class);
         $modelo = Modelo::create($request->all());
 
         return response()->json($modelo);
